+++ .gitignore
@@ -1,5 +1,6 @@
-#[Dd]ist/
+[Dd]ist/
 .vscode
+.parcel-cache
 test
 node_modules
 package/dist
