+++ .npmignore
@@ -1,0 +1,8 @@
+.parcel-cache
+.vscode
+docs
+node_modules
+src
+test
+.gitignore
+tsconfig.json
