--- package/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "@husnain.taj/konsolejs",
-  "version": "0.0.4",
-  "description": "A JavaScript Library for creating console UI on the web.",
-  "main": "dist/es5/index.js",
-  "module": "dist/es6/index.js",
-  "types": "dist/es6/index.d.ts",
-  "sideEffects": false,
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/HusnainTaj/KonsoleJS.git"
-  },
-  "keywords": [
-    "console",
-    "javascript",
-    "console",
-    "ui"
-  ],
-  "author": "Husnain Taj",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/HusnainTaj/KonsoleJS/issues"
-  },
-  "homepage": "https://github.com/HusnainTaj/KonsoleJS#readme"
-}
