+++ src/konsole.css
@@ -1,0 +1,95 @@
+:root
+{
+    --konsole-blur-opacity: 0.8;
+    --konsole-dark-color: rgb(240, 240, 240);
+    --konsole-light-color: rgb(15, 15, 15);
+
+    --konsole-color: var(--konsole-light-color);
+}
+
+.Konsole
+{
+	user-select: none;
+	opacity: var(--konsole-blur-opacity);
+	color: var(--konsole-color);
+}
+
+.Konsole *
+{
+	pointer-events: none !important;
+}
+.Konsole.dark
+{
+	--konsole-color: var(--konsole-dark-color);
+}
+.Konsole.focussed
+{
+	opacity: 1;
+}
+
+.Konsole.Konsole.focussed pre.KonsoleLine:last-of-type span.KonsoleLineText
+{
+	animation: KaretBlink 600ms steps(2) infinite;
+}
+
+.Konsole pre.KonsoleLine
+{
+	margin: 10px 0;
+	word-break: break-all;
+	white-space: pre-wrap;
+}
+.Konsole pre.KonsolePara
+{
+	margin: 5px 0;
+}
+
+.Konsole pre.KonsoleLine:last-of-type span.KonsoleLineText
+{
+	border-right: 10px solid var(--konsole-color);
+	padding-right: 2px;
+}
+
+.Konsole a
+{
+	pointer-events: all !important;
+	color: var(--konsole-color);
+	text-decoration: none;
+	border-bottom: 2px solid var(--konsole-color);
+}
+
+.Konsole a:hover
+{
+	border-bottom-style: dashed;
+}
+
+.Konsole .KonsoleChoice
+{
+	list-style-type: circle;
+}
+
+
+.Konsole .KonsoleChoice li.active
+{
+	list-style-type: disc;
+}
+
+
+#konsoleInput
+{
+    background: transparent;
+    border: none;
+    outline: none;
+    color: var(--konsole-color);
+
+    position: absolute;
+    opacity: 0;
+    left: -200%;
+    pointer-events: none;
+}
+
+@keyframes KaretBlink {
+    to
+    {
+        border:none;
+    }
+}
