+++ src/konsoleSettings.ts
@@ -1,0 +1,9 @@
+export class KonsoleSettings
+{
+    prefix: string = "$";
+    animatePrint: boolean = true;
+    printLetterInterval: number = 25;
+    registerDefaultKommands: boolean = true;
+    caseSensitiveKommands: boolean = true;
+    invalidKommandMessage: string = "invalid command.";
+}
