+++ src/main.scss
@@ -1,3 +1,5 @@
+// Replaced by konsole.css
+
 :root
 {
     --konsole-blur-opacity: 0.8;
