+++ src/parcel.d.ts
@@ -1,0 +1,4 @@
+declare module 'bundle-text:*' {
+    const value: string;
+    export default value;
+  }
