--- tsconfig.es5.json
@@ -1,11 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "commonjs",
-    "target": "ES5",
-    "declaration": true,
-    "outDir": "./package/dist/es5",
-    "moduleResolution": "node",
-    "lib": ["ES2020", "DOM"]
-  },
-  "include": ["src/**/*"]
-}
