+++ tsconfig.json
@@ -5,7 +5,12 @@
     "lib": ["ES2020", "DOM"],
     "declaration": true,
     "outDir": "./package/dist/es6",
-    "moduleResolution": "node"
+    "moduleResolution": "node",
+    // parcel
+    "baseUrl": ".",
+    "paths": {
+      "~*": ["./*"]
+    }
   },
   "include": ["src/**/*"]
 }
